+++ upsetplot/tests/test_reformat.py
@@ -1,6 +1,6 @@
 import pytest
 import pandas as pd
-from pandas.util.testing import assert_series_equal, assert_frame_equal
+from pandas.testing import assert_series_equal, assert_frame_equal
 
 from upsetplot import generate_counts, generate_samples
 from upsetplot import query